requirements.txt
PyQt5
PyOpenGL
pyqtgraph
Pillow
vmflib
numpy
pyinstaller